Clean: move some funcs to new files
